coupon 기능 세부고려사항 완료
